split android and ios modules
