+++ lib/appium_lib_core/android/uiautomator1/bridge.rb
@@ -4,7 +4,7 @@ module Appium
       module Uiautomator1
         module Bridge
           def self.for(target)
-            target.extend Appium::Android::Device
+            target.extend Appium::Core::Android::Device
             Core::Android::SearchContext.extend
           end
         end
