+++ lib/appium_lib_core/android/uiautomator2/bridge.rb
@@ -4,7 +4,7 @@ module Appium
       module Uiautomator2
         module Bridge
           def self.for(target)
-            target.extend Appium::Android::Device
+            target.extend Appium::Core::Android::Device
             Core::Android::SearchContext.extend
             target.extend Appium::Core::Android::Uiautomator2::Device
           end
